xen/arm: VGIC: wrap gic_hw_ops calls
At the moment every access to the GIC hardware, both for Xen's own
interrupt handling and for the VGIC emulation (list register
handling), goes through a structure with function pointers.
Those pointers are set up once at boot time, depending on the hardware,
and never change. However we resolve this indirection *every* time we
access the GIC, which might be costly if the CPU can't speculate this.
This is even more annyoing since some of the function just do one MMIO or
system register access.

To simplify optimising this later on, provide "static inline" wrapper
functions in a header file, and use those wrappers in the interrupt and
VGIC code.
This allows to change this later on a per function base without
cluttering the actual code.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>